refactor user.dto
